pyproject.toml
[tool.poetry]
name = "recommendation-system"
version = "0.1.0"
description = ""
authors = ["Sti_ll@mail.ru"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
flake8 = "^7.1.2"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
requests = "^2.32.3"
pillow = "^11.1.0"
ipython = "^8.32.0"
redis = "^5.2.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
coverage = "^7.6.12"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.7.0"
schedule = "^1.2.2"
gunicorn = "^23.0.0"
networkx = "^3.4.2"
numpy = "^2.2.3"
scipy = "^1.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
